Compare image times and sample times as signed ints
The problem that we were comparing signed and unsigned ints was originally
found by Lubomir. As image times can be negative (and the comparison should
fail in that case) we need to compared them as signed.

Acked-by: Lubomir I. Ivanov <neolit123@gmail.com>

+++ qt-ui/divelogexportdialog.cpp
@@ -334,7 +334,7 @@ void DiveLogExportDialog::export_depths(const char *filename, const bool selecte
 			int n = dive->dc.samples;
 			struct sample *s = dive->dc.sample;
 			depth.mm = 0;
-			while (--n >= 0 && s->time.seconds <= picture->offset.seconds) {
+			while (--n >= 0 && (int32_t)s->time.seconds <= picture->offset.seconds) {
 				depth.mm = s->depth.mm;
 				s++;
 			}
